Updated common events
- Added common logger
