--- src/main/java/com/minecrafttas/common/events/EventBase.java
@@ -1,5 +1,0 @@
-package com.minecrafttas.common.events;
-
-public interface EventBase {
-
-}
