+++ src/main/java/com/minecrafttas/common/events/EventListener.java
@@ -19,4 +19,8 @@ public class EventListener {
 	public static ArrayList<EventBase> getEventListeners(){
 		return EVENTLISTENER_REGISTRY;
 	}
+	
+	public static interface EventBase {
+
+	}
 }
