--- src/main/java/com/minecrafttas/common/events/server/EventServerGameLoop.java
@@ -1,20 +1,0 @@
-package com.minecrafttas.common.events.server;
-
-import com.minecrafttas.common.events.EventBase;
-import com.minecrafttas.common.events.EventListener;
-
-import net.minecraft.server.MinecraftServer;
-
-public interface EventServerGameLoop extends EventBase {
-
-	public void onRunServerGameLoop(MinecraftServer server);
-	
-	public static void fireOnServerGameLoop(MinecraftServer server) {
-		for (EventBase eventListener : EventListener.getEventListeners()) {
-			if(eventListener instanceof EventServerGameLoop) {
-				EventServerGameLoop event = (EventServerGameLoop) eventListener;
-				event.onRunServerGameLoop(server);
-			}
-		}
-	}
-}
