+++ src/main/java/com/minecrafttas/tasmod/events/server/EventLoadstate.java
@@ -2,7 +2,7 @@ package com.minecrafttas.tasmod.events.server;
 
 import java.io.File;
 
-import com.minecrafttas.common.events.EventBase;
+import com.minecrafttas.common.events.EventListener.EventBase;
 import com.minecrafttas.tasmod.TASmod;
 import com.minecrafttas.tasmod.events.TASmodEventListener;
 import com.minecrafttas.tasmod.util.LoggerMarkers;
