+++ src/main/java/com/minecrafttas/tasmod/savestates/server/chunkloading/SavestatesChunkControl.java
@@ -5,8 +5,8 @@ import java.util.List;
 import com.minecrafttas.tasmod.TASmod;
 import com.minecrafttas.tasmod.mixin.savestates.MixinChunkProviderClient;
 import com.minecrafttas.tasmod.mixin.savestates.MixinChunkProviderServer;
-import com.minecrafttas.tasmod.util.LoggerMarkers;
 import com.minecrafttas.tasmod.util.Ducks.ChunkProviderDuck;
+import com.minecrafttas.tasmod.util.LoggerMarkers;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
