+++ src/main/java/com/minecrafttas/tasmod/util/ShieldDownloader.java
@@ -17,8 +17,8 @@ import java.util.Map;
 import javax.net.ssl.HttpsURLConnection;
 
 import com.google.common.collect.Maps;
-import com.minecrafttas.common.events.client.player.EventOtherPlayerJoinedClientSide;
-import com.minecrafttas.common.events.client.player.EventPlayerJoinedClientSide;
+import com.minecrafttas.common.events.EventClient.EventOtherPlayerJoinedClientSide;
+import com.minecrafttas.common.events.EventClient.EventPlayerJoinedClientSide;
 import com.mojang.authlib.GameProfile;
 
 import net.minecraft.client.Minecraft;
